requirements.txt
flask
flask_login
flask-sqlalchemy
flask-migrate
psycopg2-binary
python-dotenv
requests
pytest-cov